+++ pr1-coursework/ChopStep.java
@@ -1,3 +1,4 @@
+import greenfoot.Greenfoot;
 import greenfoot.GreenfootImage;
 
 public class ChopStep extends Step{
@@ -7,14 +8,30 @@ public class ChopStep extends Step{
     
 
     public ChopStep(Location location, String name, int chopsRequired){
-        super(location); 
+        super(location, name); 
         this.chopsRequired = chopsRequired;
         this.chopCount = 0;
         this.cDown = false;
         setImage(name + ".png");
     }
+        
 
-    public boolean chop(){
+    private void checkForChop(){
+        if (Greenfoot.isKeyDown("c") && !cDown){
+            chop();
+            cDown = true;
+        }
+        if (!Greenfoot.isKeyDown("c") && cDown){
+            cDown = false;
+        }
+    }
+    
+    private void chop(){
+        chopCount++;
+        if (chopCount >= chopsRequired){
+            setImage(ingredientName + "-chopped.png");
+            setStepComplete(true);
+        }
         
     }
 }
