--- pr1-coursework/CookType.java
@@ -1,3 +1,0 @@
-public enum CookType {
-    BOIL,FRY,OVEN_BAKE
-}
