--- pr1-coursework/CookableIngredient.ctxt
@@ -1,11 +1,0 @@
-#BlueJ class context
-comment0.target=CookableIngredient
-comment0.text=\n\ Write\ a\ description\ of\ class\ CookableIngredient\ here.\n\ \n\ @author\ (your\ name)\ \n\ @version\ (a\ version\ number\ or\ a\ date)\n
-comment1.params=ingredientNumber\ name\ chopsRequired\ cookMethod\ cookTime\ burnTime
-comment1.target=CookableIngredient(int,\ java.lang.String,\ int,\ CookType,\ int,\ int)
-comment2.params=
-comment2.target=void\ onAct()
-comment2.text=\n\ Act\ -\ do\ whatever\ the\ CookableIngredient\ wants\ to\ do.\ This\ method\ is\ called\ whenever\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\n
-comment3.params=
-comment3.target=void\ cook()
-numComments=4
