+++ pr1-coursework/Ingredient.java
@@ -13,31 +13,14 @@ import java.util.ArrayList;
  * @version (a version number or a date)
  */
 public class Ingredient extends Actor {
-    private int ingredientNumber;
-    private String name;
-    protected ArrayList<GreenfootImage> icons;
-    private int chopsRequired;
-    private int chops;
-    private boolean cDown;
-    public Location location ;
-    protected int progress;
+    private ArrayList<Step> steps;
+    private int currentStep;
+    public Location myLocation ;
     public boolean isPrepared;
 
     
-    public Ingredient(int ingredientNumber, String name, int chopsRequired) {
-        this.ingredientNumber = ingredientNumber;
-        this.name = name;
-        this.icons = new ArrayList<GreenfootImage>();
-        this.icons.add(new GreenfootImage(name + ".png"));
-        this.icons.add(new GreenfootImage(name + "-chopped.png"));
-        this.chopsRequired = chopsRequired;
-        this.chops = 0;
-        this.cDown = false;
-        this.location = Location.PREP_AREA;
-        this.isPrepared = false;
-        this.progress = 0; 
-        
-        setImage(icons.get(progress));
+    public Ingredient() {
+    
 
     }
 
@@ -47,38 +30,14 @@ public class Ingredient extends Actor {
      */
     public void act(){ 
     
-       if (location == Location.PREP_AREA && progress < 1){
-           checkForChop();
-       }
-       onAct();
-       if (progress == icons.size()-1) {
-           isPrepared = true;
-       }
-    }
-    
-    protected void onAct(){
+    //    if (location == ){
+    //        checkForChop();
+    //    }
+    //    onAct();
+    //    if (progress == icons.size()-1) {
+    //        isPrepared = true;
+    //    }
     }
 
-    private void checkForChop(){
-        if (Greenfoot.isKeyDown("c") && cDown == false){
-            chop();
-            cDown = true;
-        }
-        else if (!Greenfoot.isKeyDown("c") && cDown == true){
-            cDown = false;
-        }
-    }
-    
-    private void chop(){
-        chops++;
-        if (chops >= chopsRequired){
-            incrementProgress();
-        }
-        
-    }
-    
-    private void incrementProgress(){
-        progress++;
-        setImage(icons.get(progress));
-    }
+
 }
