+++ pr1-coursework/project.greenfoot
@@ -7,18 +7,15 @@ dependency1.type=UsesDependency
 dependency2.from=MyWorld
 dependency2.to=CookableIngredient
 dependency2.type=UsesDependency
-dependency3.from=MyWorld
+dependency3.from=CookableIngredient
 dependency3.to=CookType
 dependency3.type=UsesDependency
-dependency4.from=CookableIngredient
-dependency4.to=CookType
-dependency4.type=UsesDependency
 editor.fx.0.height=727
 editor.fx.0.width=1284
 editor.fx.0.x=69
 editor.fx.0.y=113
 height=716
-package.numDependencies=4
+package.numDependencies=3
 package.numTargets=4
 project.charset=UTF-8
 publish.hasSource=false
@@ -65,5 +62,5 @@ target4.y=0
 version=3.1.0
 width=1062
 world.lastInstantiated=MyWorld
-xPosition=42
-yPosition=97
+xPosition=181
+yPosition=81
